# openapi.yaml
openapi: 3.0.3
info:
  title: INDULINK E-Commerce Platform API
  description: Comprehensive API for the INDULINK e-commerce platform supporting customers, suppliers, and administrators
  version: 1.0.0
  contact:
    email: api-support@indulink.com
    url: https://docs.indulink.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.indulink.com/api
    description: Production server
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get all products
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/categoryParam'
        - $ref: '#/components/parameters/searchParam'
        - $ref: '#/components/parameters/minPriceParam'
        - $ref: '#/components/parameters/maxPriceParam'
        - $ref: '#/components/parameters/sortParam'
        - $ref: '#/components/parameters/orderParam'
      responses:
        '200':
          description: Products retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'

    post:
      tags:
        - Products
      summary: Create product (Supplier/Admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
...
  /orders:
    post:
      tags:
        - Orders
      summary: Create order (Customer)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'

    get:
      tags:
        - Orders
      summary: Get customer orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Orders retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: Get all users (Admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Users retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication using Bearer token
  parameters:
...
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Page number
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 10
      description: Items per page
    categoryParam:
      name: category
      in: query
      schema:
        type: string
      description: Filter by category
    searchParam:
      name: search
      in: query
      schema:
        type: string
      description: Search term
    minPriceParam:
      name: minPrice
      in: query
      schema:
        type: number
      description: Minimum price filter
    maxPriceParam:
      name: maxPrice
      in: query
      schema:
        type: number
      description: Maximum price filter
    sortParam:
      name: sort
      in: query
      schema:
        type: string
        enum: [name, price, createdAt, rating]
      description: Sort field
    orderParam:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
      description: Sort order
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          example: 5f8d0d55b54764421b7156a1
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        phone:
          type: string
          example: +1234567890
        role:
          type: string
          enum: [customer, supplier, admin]
          example: customer
        createdAt:
          type: string
          format: date-time
          example: 2023-10-15T10:30:00.000Z
...
    Product:
      type: object
      properties:
        id:
          type: string
          example: 5f8d0d55b54764421b7156b2
        name:
          type: string
          example: Premium Headphones
        description:
          type: string
          example: High-quality wireless headphones
        price:
          type: number
          format: float
          example: 199.99
        category:
          type: string
          example: electronics
        stock:
          type: integer
          example: 50
        sku:
          type: string
          example: HP-001
        images:
          type: array
          items:
            type: string
          example: ["headphones.jpg"]
        rating:
          type: number
          format: float
          example: 4.5
        reviewsCount:
          type: integer
          example: 12
        createdAt:
          type: string
          format: date-time
          example: 2023-10-10T09:00:00.000Z
    ProductResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            product:
              $ref: '#/components/schemas/Product'
    ProductListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            products:
              type: array
              items:
                $ref: '#/components/schemas/Product'
            pagination:
              $ref: '#/components/schemas/Pagination'
    ProductCreate:
      type: object
      required:
        - name
        - price
        - category
      properties:
        name:
          type: string
          example: New Product
        description:
          type: string
          example: Product description
        price:
          type: number
          format: float
          example: 99.99
        category:
          type: string
          example: electronics
        stock:
          type: integer
          example: 100
        sku:
          type: string
          example: NEW-001
        images:
          type: array
          items:
            type: string
          example: ["product.jpg"]
...
    Order:
      type: object
      properties:
        id:
          type: string
          example: ord12345
        userId:
          type: string
          example: 5f8d0d55b54764421b7156a1
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled, refunded]
          example: pending
        totalAmount:
          type: number
          format: float
          example: 399.98
        shippingAddress:
          type: string
          example: 123 Main St, City, Country
        paymentMethod:
          type: string
          example: credit_card
        createdAt:
          type: string
          format: date-time
          example: 2023-10-15T12:00:00.000Z
    OrderItem:
      type: object
      properties:
        productId:
          type: string
          example: 5f8d0d55b54764421b7156b2
        quantity:
          type: integer
          example: 2
        price:
          type: number
          format: float
          example: 199.99
        name:
          type: string
          example: Premium Headphones
    OrderResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Order created successfully
        data:
          type: object
          properties:
            order:
              $ref: '#/components/schemas/Order'
    OrderListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            orders:
              type: array
              items:
                $ref: '#/components/schemas/Order'
    OrderCreate:
      type: object
      required:
        - items
        - totalAmount
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        shippingAddress:
          type: string
          example: 123 Main St, City, Country
        paymentMethod:
          type: string
          example: credit_card
        totalAmount:
          type: number
          format: float
          example: 399.98
        shippingMethod:
          type: string
          example: standard

    # Category Schemas
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        totalPages:
          type: integer
          example: 10
...
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            users:
              type: array
              items:
                $ref: '#/components/schemas/User'
            total:
              type: integer
              example: 100